requirements.txt
Django==3.2
Pillow==8.2.0
python-telegram-bot==13.4.1
psycopg2-binary==2.8.6
python-dotenv==0.17.0
djangorestframework==3.12.4
markdown==3.3.4
django-filter==2.4.0
